Module Kv_hash__Frontend

The frontend: record updates in two logs (rotating between them).

module Nv_map_ss_ = Kv_hash.Nv_map_ss_private.Nv_map_ss0

module Merge_process_ : sig ... end

module KV : sig ... end

module Op : sig ... end

module Log_file_w : sig ... end

module Log_file_r : sig ... end

module Control_fields : sig ... end

module Control : sig ... end

module Writer_1 : sig ... end

The writer is responsible for taking updates and recording in log, and periodically rotating logs and firing the merge process.

module type WRITER = sig ... end

module Writer_2 : WRITER

module Writer = Writer_2

module Test : functor () sig ... end